Remove unused extended classes
